website/requirements.txt
django==5.1
django-debug-toolbar==4.4.6
django-imagekit==5.0.0
django-random-queryset==0.2.0
django-storages==1.14.4
boto3==1.34.158
pillow==9.5.0
django-unfold==0.35.0
django_browser_reload==1.13.0
django-pipeline==3.1.0
python-resize-image==1.1.20
python-decouple==3.8
psycopg2==2.9.9